Change controllers to authorize and add fund totals

+++ ChurchGivingRecorder/Models/Deposit.cs
@@ -21,6 +21,6 @@ namespace ChurchGivingRecorder.Models
 
         [NotMapped]
         [DisplayFormat(DataFormatString = "{0:c}")]
-        public double TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
     }
 }
